# docs/openapi.yaml
openapi: 3.1.0
info:
  description: |-
    The PaymentFlow payments API. This document describes the public `/v1` surface — the endpoints PaymentFlow versions, deprecates on a published timeline, and supports for integrators. Endpoints not described here are internal and carry no compatibility promise.

    Amounts are integers in the currency's minor unit (`amountMinor`): `1000` in `USD` is $10.00. There are no floating-point amounts anywhere in this API.

    Every mutating request accepts an `Idempotency-Key` header and every list is cursor-paginated; both are described on the operations that use them.
  license:
    name: Proprietary
  title: PaymentFlow API
  version: 2026-08-01
servers:
- url: https://api.paymentflow.dev
  description: PaymentFlow API
security:
- SecretKey: []
paths:
...
  /v1/webhook_endpoints:
    get:
      description: "Returns every endpoint registered for your merchant in the mode of the key you called with, including any the platform has auto-disabled. Not paginated: the number of endpoints per merchant and mode is capped, so this is always one short list."
      operationId: listWebhookEndpoints
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WebhookEndpointResponse"
          description: Your webhook endpoints.
        "401":
          content:
            application/json:
              example:
                code: UNAUTHORIZED
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#unauthorized
                message: Authentication is required or has failed.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 401
                timestamp: 2026-07-28T12:34:56.789Z
                type: authentication_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Authentication is required or has failed.
        "403":
          content:
            application/json:
              example:
                code: INSUFFICIENT_SCOPE
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#insufficient_scope
                message: This API key does not have the required scope for this endpoint.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 403
                timestamp: 2026-07-28T12:34:56.789Z
                type: permission_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: This API key does not have the required scope for this endpoint.
        "429":
          content:
            application/json:
              example:
                code: RATE_LIMIT_EXCEEDED
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#rate_limit_exceeded
                message: Too many requests. Slow down and retry after the interval in the Retry-After header.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 429
                timestamp: 2026-07-28T12:34:56.789Z
                type: rate_limit_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Too many requests. Slow down and retry after the interval in the Retry-After header.
        "500":
          content:
            application/json:
              example:
                code: INTERNAL_ERROR
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#internal_error
                message: An unexpected error occurred.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 500
                timestamp: 2026-07-28T12:34:56.789Z
                type: api_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: An unexpected error occurred.
      summary: List your webhook endpoints
      tags:
      - Webhook endpoints
    post:
      description: |-
        Registers a URL to receive events, and returns the signing secret.

        **The signing secret is shown exactly once, here.** It cannot be retrieved afterwards — only a hash is kept — so store it before you do anything else. Every delivery to this endpoint is signed with it, and verifying that signature is what distinguishes a genuine delivery from anyone who has learned your URL.

        The URL must be reachable over HTTPS from the public internet. Private, link-local and cloud-metadata addresses are refused at registration and re-checked at delivery time, so a hostname cannot be repointed at one later.
      operationId: createWebhookEndpoint
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWebhookEndpointRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEndpointCreatedResponse"
          description: The endpoint is registered. This response carries the signing secret; it is not shown again.
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
          description: "The endpoint was refused: a URL that is not HTTPS, resolves to a private or metadata address, or an empty `enabledEvents`."
        "401":
          content:
            application/json:
              example:
                code: UNAUTHORIZED
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#unauthorized
                message: Authentication is required or has failed.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 401
                timestamp: 2026-07-28T12:34:56.789Z
                type: authentication_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Authentication is required or has failed.
        "403":
          content:
            application/json:
              example:
                code: INSUFFICIENT_SCOPE
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#insufficient_scope
                message: This API key does not have the required scope for this endpoint.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 403
                timestamp: 2026-07-28T12:34:56.789Z
                type: permission_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: This API key does not have the required scope for this endpoint.
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
          description: You already have an endpoint registered at this URL in this mode.
        "429":
          content:
            application/json:
              example:
                code: RATE_LIMIT_EXCEEDED
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#rate_limit_exceeded
                message: Too many requests. Slow down and retry after the interval in the Retry-After header.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 429
                timestamp: 2026-07-28T12:34:56.789Z
                type: rate_limit_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Too many requests. Slow down and retry after the interval in the Retry-After header.
        "500":
          content:
            application/json:
              example:
                code: INTERNAL_ERROR
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#internal_error
                message: An unexpected error occurred.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 500
                timestamp: 2026-07-28T12:34:56.789Z
                type: api_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: An unexpected error occurred.
      summary: Register a webhook endpoint
      tags:
      - Webhook endpoints
...
  /v1/webhook_endpoints/{id}/rotate_secret:
    post:
      description: |-
        Issues a new signing secret for the endpoint and returns it — the second and only other place a raw secret is ever shown.

        **Both secrets verify during a grace window**, so you can deploy the new one without dropping in-flight deliveries: rotate, deploy, and let the old secret lapse. Verifying against either during that window is the intended behaviour, not a weakness.

        Rotate when a secret may have been exposed, or when you have lost the one you were shown at registration — it cannot be recovered, only replaced.
      operationId: rotateWebhookEndpointSecret
      parameters:
      - description: The endpoint whose secret to rotate.
        in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEndpointCreatedResponse"
          description: "The endpoint, with its new signing secret. Not shown again."
        "401":
          content:
            application/json:
              example:
                code: UNAUTHORIZED
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#unauthorized
                message: Authentication is required or has failed.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 401
                timestamp: 2026-07-28T12:34:56.789Z
                type: authentication_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Authentication is required or has failed.
        "403":
          content:
            application/json:
              example:
                code: INSUFFICIENT_SCOPE
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#insufficient_scope
                message: This API key does not have the required scope for this endpoint.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 403
                timestamp: 2026-07-28T12:34:56.789Z
                type: permission_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: This API key does not have the required scope for this endpoint.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
          description: "No such endpoint. Also returned when it exists but belongs to another merchant, or to the other mode — never `403`, which would confirm it exists."
        "429":
          content:
            application/json:
              example:
                code: RATE_LIMIT_EXCEEDED
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#rate_limit_exceeded
                message: Too many requests. Slow down and retry after the interval in the Retry-After header.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 429
                timestamp: 2026-07-28T12:34:56.789Z
                type: rate_limit_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Too many requests. Slow down and retry after the interval in the Retry-After header.
        "500":
          content:
            application/json:
              example:
                code: INTERNAL_ERROR
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#internal_error
                message: An unexpected error occurred.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 500
                timestamp: 2026-07-28T12:34:56.789Z
                type: api_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: An unexpected error occurred.
      summary: Rotate an endpoint's signing secret
      tags:
      - Webhook endpoints
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        code:
          type: string
          description: "A stable, machine-readable identifier for this specific failure, such as `PAYMENT_NOT_CAPTURABLE`. The set of codes grows by policy, which is why `type` and not this is what a client should switch on. Every code is catalogued in docs/ERRORS.md."
        correlationId:
          type: string
          description: "Identifies the whole distributed trace this call belongs to, which may span several services. Broader than `requestId`."
        docUrl:
          type: string
          description: Where to read about this specific code.
        errors:
          type: array
          description: "Field-level validation failures, when the request failed validation in more than one place."
          items:
            $ref: "#/components/schemas/ApiFieldError"
        message:
          type: string
          description: "A human-readable explanation. Safe to log; not intended to be shown to your customers, and never programmatically parsed."
        param:
          type: string
          description: "The single offending parameter, when there is exactly one. Multi-field validation failures use `errors` instead."
        path:
          type: string
          description: The request path that produced the error.
        requestId:
          type: string
          description: Identifies this one HTTP call. **Quote this in a support request** — it is on every log line and every row of your request log.
        status:
          type: integer
          format: int32
          description: "The HTTP status code, repeated in the body so a logged response is self-contained."
        timestamp:
          type: string
          format: date-time
          description: "When the error occurred, as RFC 3339."
        type:
          type: string
          description: "The error's classification, and **the field to branch on**. This is a small closed set an SDK maps to exception classes: `authentication_error`, `permission_error`, `invalid_request_error`, `idempotency_error`, `rate_limit_error`, `api_error`."
    ApiFieldError:
      type: object
      properties:
        field:
          type: string
          description: The request field that failed validation.
        message:
          type: string
          description: Why it failed. The rejected value is deliberately never echoed back.
...
    CreateWebhookEndpointRequest:
      type: object
      properties:
        description:
          type: string
          description: Your own label for this endpoint. Useful once you have more than one.
          example: Production order pipeline
          maxLength: 255
          minLength: 0
        enabledEvents:
          type: array
          description: "The event types to send here. Must name at least one — an endpoint subscribed to nothing looks identical to a broken platform from your side, so it is refused rather than accepted silently. Use `[\"*\"]` to receive everything, including types added later."
          example:
          - payment.captured
          - payment.refunded
          items:
            type: string
            maxLength: 64
            minLength: 0
        metadata:
          type: object
          additionalProperties:
            type: string
          description: "Up to 20 arbitrary key-value pairs you can attach to this object. PaymentFlow never interprets them; they are returned on every read and are filterable with `metadata[key]=value`."
        url:
          type: string
          description: |-
            Where to send deliveries. Must be reachable from the public internet over HTTPS: private, link-local and cloud-metadata addresses are refused, and the address is re-checked at delivery time so a hostname cannot be repointed at one afterwards.

            Not updatable later — the URL is half of an endpoint's identity.
          example: https://example.com/webhooks/paymentflow
          maxLength: 2048
          minLength: 0
      required:
      - enabledEvents
      - url
...
    WebhookEndpointResponse:
      type: object
      properties:
        apiVersion:
          type: string
          description: The API revision delivery bodies are rendered in for this endpoint.
          example: 2026-08-01
        consecutiveFailureCount:
          type: integer
          format: int32
          description: How many deliveries have failed in a row. Resets to zero on the first success. When it reaches the platform's threshold the endpoint is disabled automatically — a permanently broken receiver is not retried forever.
          example: 0
        createdAt:
          type: string
          format: date-time
          description: "When the endpoint was registered, as RFC 3339."
        description:
          type: string
          description: Your own label for this endpoint.
          example: Production order pipeline
        disabledAt:
          type: string
          format: date-time
          description: "When the endpoint was disabled, if it is."
        disabledReason:
          type: string
          description: Why the endpoint was disabled. **This is how you tell "the platform turned this off because it kept failing" from "I turned this off"** — which decides whether re-enabling it requires fixing something first. Absent on an enabled endpoint.
          enum:
          - CONSECUTIVE_FAILURES
        enabled:
          type: boolean
          description: Whether deliveries are currently being sent. May be false because you disabled it or because the platform did — see `disabledReason`.
        enabledEvents:
          type: array
          description: "The event types this endpoint receives. `*` means every type, including ones added later."
          example:
          - payment.captured
          - payment.refunded
          items:
            type: string
        id:
          type: string
          format: uuid
          description: Unique identifier for this endpoint.
        metadata:
          type: object
          additionalProperties:
            type: string
          description: "Up to 20 arbitrary key-value pairs you can attach to this object. PaymentFlow never interprets them; they are returned on every read and are filterable with `metadata[key]=value`."
        migratedFromLegacy:
          type: boolean
          description: Whether this endpoint was carried over from the platform's pre-webhook-product configuration.
        object:
          type: string
          description: Always `webhook_endpoint`.
          example: webhook_endpoint
        signingSecretPrefix:
          type: string
          description: "The first few characters of the signing secret, so you can tell which secret an endpoint holds. **The full secret is shown exactly once**, when the endpoint is created or its secret is rotated, and is not retrievable afterwards."
          example: whsec_9f2c
        updatedAt:
          type: string
          format: date-time
          description: "When the endpoint last changed, as RFC 3339."
        url:
          type: string
          description: "Where deliveries are sent. Not updatable: the URL is half of an endpoint's identity, and repointing one would leave its delivery history attached to a destination that never received any of it. Register a new endpoint instead."
          example: https://example.com/webhooks/paymentflow
  securitySchemes:
    SecretKey:
      bearerFormat: sk
      description: |-
        Your secret API key, sent as `Authorization: Bearer sk_test_...`.

        The key determines both *whose* data you see and *which mode* you see it in: an `sk_test_` key reads and writes test data only, an `sk_live_` key live data only. No header, parameter, or body field can change either — to work in the other mode, use the other key.
      scheme: bearer
      type: http